requirements.txt
Django~=4.2.2
python-dotenv~=1.0.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pillow==10.4.0
